auto-kj/requirements.txt
python-mpv>=1.0.7
yt-dlp  # updated to latest on every startup
spleeter>=2.4.0
openwakeword>=0.6.0
openai-whisper>=20231117
anthropic>=0.40.0
evdev>=1.7.0
JACK-Client>=0.5.4
numpy>=1.24.0
requests>=2.28.0
